Demonstration of standalone LunaSteps.
* No Lua involved in Steps.
* New macros for simple cpp file.

As before, comments/critiques are welcome.
