+++ .gitignore
@@ -128,3 +128,4 @@ bundle/ffmpeg/*/*.def
 bundle/ffmpeg/*/*.orig.def
 bundle/ffmpeg/*/*.dll.a
 bundle/ffmpeg/*/*.lib
+*.xcbkptlist
