+++ src/StepMania-net2011.vcxproj
@@ -507,6 +507,7 @@ cl /Zl /nologo /c verstub.cpp /Fo"$(IntDir)"</Command>
     <ClCompile Include="SoundEffectControl.cpp" />
     <ClCompile Include="StageStats.cpp" />
     <ClCompile Include="Steps.cpp" />
+    <ClCompile Include="LunaSteps.cpp" />
     <ClCompile Include="StepsUtil.cpp" />
     <ClCompile Include="Style.cpp" />
     <ClCompile Include="StyleUtil.cpp" />
@@ -1822,6 +1823,7 @@ cl /Zl /nologo /c verstub.cpp /Fo"$(IntDir)"</Command>
     <ClInclude Include="SoundEffectControl.h" />
     <ClInclude Include="StageStats.h" />
     <ClInclude Include="Steps.h" />
+    <ClInclude Include="LunaSteps.h" />
     <ClInclude Include="StepsUtil.h" />
     <ClInclude Include="Style.h" />
     <ClInclude Include="StyleUtil.h" />
