+++ src/StepMania-net2013.vcxproj
@@ -506,6 +506,7 @@ cl /Zl /nologo /c verstub.cpp /Fo"$(IntDir)"</Command>
     <ClCompile Include="SoundEffectControl.cpp" />
     <ClCompile Include="StageStats.cpp" />
     <ClCompile Include="Steps.cpp" />
+    <ClCompile Include="LunaSteps.cpp" />
     <ClCompile Include="StepsUtil.cpp" />
     <ClCompile Include="Style.cpp" />
     <ClCompile Include="StyleUtil.cpp" />
@@ -1821,6 +1822,7 @@ cl /Zl /nologo /c verstub.cpp /Fo"$(IntDir)"</Command>
     <ClInclude Include="SoundEffectControl.h" />
     <ClInclude Include="StageStats.h" />
     <ClInclude Include="Steps.h" />
+    <ClInclude Include="LunaSteps.h" />
     <ClInclude Include="StepsUtil.h" />
     <ClInclude Include="Style.h" />
     <ClInclude Include="StyleUtil.h" />
