+++ src/StepMania-net2013.vcxproj.filters
@@ -495,6 +495,9 @@
     <ClCompile Include="Steps.cpp">
       <Filter>Data Structures</Filter>
     </ClCompile>
+    <ClCompile Include="LunaSteps.cpp">
+      <Filter>Data Structures</Filter>
+    </ClCompile>
     <ClCompile Include="StepsUtil.cpp">
       <Filter>Data Structures</Filter>
     </ClCompile>
@@ -1998,6 +2001,9 @@
       <Filter>Data Structures</Filter>
     </ClInclude>
     <ClInclude Include="Steps.h">
+      <Filter>Data Structures</Filter>
+    </ClInclude>
+    <ClInclude Include="LunaSteps.h">
       <Filter>Data Structures</Filter>
     </ClInclude>
     <ClInclude Include="StepsUtil.h">
